Change over to HTMLeditior, Fixed save and title bugs
Changed from textArea to HTMLeditior to allow text formating like in
word

Fixed bug where the save button would not use the file location gained
by opening a file and try and get the user to use a save-as operation.


Fixed a bug where when a document would be loaded and then added its
name next the previous documents loaded name.